refactored edit post modal
