+++ components/account-page/AccountPostCard.tsx
@@ -1,12 +1,9 @@
-import { Button, Grid, GridItem, Icon, useDisclosure } from "@chakra-ui/react";
-import EditPostModal from "../EditPostModal";
-import { v4 as uuid } from "uuid";
-import { useState } from "react";
+import { Grid, GridItem } from "@chakra-ui/react";
 import PostCard from "../PostCard";
 import { Post } from "@prisma/client";
 import { usePostQuery } from "@lib/hooks/post";
-import { PencilAltIcon } from "@heroicons/react/outline";
 import DeletePostForm from "@components/post-page/DeletePostForm";
+import EditPostForm from "@components/edit-post/EditPostForm";
 
 type AccountPostCardProps = {
   post: Post;
@@ -19,15 +16,6 @@ export default function AccountPostCard({
 }: AccountPostCardProps) {
   const { data: postsWithBookWithUser } = usePostQuery(post.id);
 
-  const { isOpen, onClose, onOpen } = useDisclosure();
-
-  const [editPostModalKey, setEditPostModalKey] = useState("");
-  const handleEditPostClick = () => {
-    // Reset edit post modal each time it is opened
-    setEditPostModalKey(uuid());
-    onOpen();
-  };
-
   const card = postsWithBookWithUser ? (
     <>
       <Grid templateColumns="repeat(2, 1fr)" rowGap={1} columnGap={1}>
@@ -35,25 +23,14 @@ export default function AccountPostCard({
           <PostCard post={postsWithBookWithUser} isLinkActive={isLinkActive} />
         </GridItem>
         <GridItem colSpan={1}>
-          <Button
-            colorScheme={"teal"}
-            type={"button"}
-            onClick={handleEditPostClick}
-            leftIcon={<Icon as={PencilAltIcon} />}
+          <EditPostForm
+            post={postsWithBookWithUser}
             width="100%"
-          >
-            Edit
-          </Button>
+          ></EditPostForm>
         </GridItem>
         <GridItem colSpan={1}>
           <DeletePostForm post={post} isAccountPage={true} width="100%" />
         </GridItem>
-        <EditPostModal
-          key={editPostModalKey}
-          isOpen={isOpen}
-          onClose={onClose}
-          post={postsWithBookWithUser}
-        />
       </Grid>
     </>
   ) : (
