+++ components/edit-post/EditPostModal.tsx
@@ -15,9 +15,9 @@ import { FormEvent, useState } from "react";
 import { useEditPostMutation } from "@lib/hooks/post";
 import { getFloatStringPriceAsNumber } from "@lib/helpers/priceHelpers";
 import { PostWithBook, PostWithBookWithUser } from "@lib/services/post";
-import UploadTextbookCover from "./create-post-page/UploadTextbookCover";
-import ViewTextbookCover from "./create-post-page/ViewTextbookCover";
-import EditPriceAndDescription from "./EditPriceAndDescription";
+import UploadTextbookCover from "../create-post-page/UploadTextbookCover";
+import ViewTextbookCover from "../create-post-page/ViewTextbookCover";
+import EditPriceAndDescription from "../EditPriceAndDescription";
 import { handleRawImage } from "@lib/helpers/frontend/handle-raw-image";
 
 export interface EditPostProps {
@@ -83,7 +83,6 @@ export default function EditPostModal({
     <Modal
       isOpen={!mutation.isSuccess && isOpen}
       onClose={onClose ?? (() => {})}
-      size="xl"
     >
       <ModalOverlay />
       <ModalContent pt={5} pb={4}>
